Move run utils to new files
